Get the raw files of datasets from the Sustentarea's Global Syndemic Project

get_raw_data() downloads the raw datasets from the Sustentarea's Global Syndemic Project.

Click here or use data(package = "gsdata") to see all datasets available.

Usage

get_raw_data(data, dir)

Arguments

data

A character string with the name of the dataset to be retrieved.

dir

A character string with the directory where the raw data file will be saved.

Value

A character string with the file path of the raw data.

See also

Examples

if (interactive()) {
  get_raw_data("gini", tempdir())
}